+++ Sources/DiceKit/Dice.swift
@@ -3,7 +3,7 @@
 /// The properties of `Dice` objects are immutable; use the addition operators to combine them with other `Die` objects or modifiers. You can use compound assignment operators if you want, so long as you declare the `Dice` object as a `var` instead of a `let` constant.
 ///
 /// - Author: Samasaur
-public struct Dice {
+public struct Dice: Caching {
     /// The dice that make up this collection, along with how many times they appear.
     ///
     /// This `[Die: Int]` dictionary stores the types of dice that appear, paired with the number of times they appear. For example:
@@ -266,9 +266,39 @@ public struct Dice {
 
     /// The probabilities of all possible rolls.
     ///
+    ///  Since 0.22.0, caches previous computations, even if they were on different objects.
+    ///  See `enableCaching`, `ENABLE_CACHING` for caching configuration
+    ///
     /// - Since: 0.17.0
     public var probabilities: Chances {
-        return __probabilities.value(input: self)
+        if let val = Dice.__cache?[self] {
+            return val
+        }
+        let val = __probabilities.value(input: self)
+        Dice.__cache?[self] = val
+        return val
+    }
+
+    fileprivate static var __cache: [Dice: Chances]? = [:]
+
+    /// Whether or not `Dice` should cache the results of probability computations across objects.
+    ///
+    /// **Note:** The results of rolling are **NOT** cached.
+    ///
+    /// Setting this value to `false` and then to `true` will clear the cache.
+    /// See `ENABLE_CACHING` for configuration of caching for all types at once.
+    ///
+    /// - Since: 0.22.0
+    public static var enableCaching = true {
+        didSet {
+            if enableCaching == false {
+                __cache = nil
+            } else {
+                if __cache == nil {
+                    __cache = [:]
+                }
+            }
+        }
     }
 }
 
@@ -418,6 +448,13 @@ extension Dice: Equatable {
             return false
         }
         return true
+    }
+}
+
+extension Dice: Hashable {
+    public func hash(into hasher: inout Hasher) {
+        hasher.combine(self.dice)
+        hasher.combine(self.modifier)
     }
 }