+++ Tests/LinuxMain.swift
@@ -6,4 +6,3 @@ var tests = [XCTestCaseEntry]()
 tests += DiceKitTests.__allTests()
 
 XCTMain(tests)
-